services/python/knowledge-service/pyproject.toml
[tool.poetry]
name = "knowledge-service"
version = "0.1.0"
description = "AI 智能投标系统 — 企业知识库服务"
authors = ["BidAI Team <dev@bidai.local>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = {extras = ["standard"], version = "^0.29"}
pydantic = "^2.0"
pydantic-settings = "^2.0"
sqlalchemy = {extras = ["asyncpg"], version = "^2.0"}
alembic = "^1.13"
celery = "^5.3"
redis = "^5.0"
httpx = "^0.27"
python-multipart = "^0.0.9"
python-json-logger = "^2.0"
minio = "^7.2"
pymupdf = "^1.23"
python-docx = "^1.1"
pillow = "^10.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4"
pytest-asyncio = "^0.23"
pytest-cov = "^4.1"
mypy = "^1.9"
ruff = "^0.4"
black = "^24.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py311']

[tool.ruff]
line-length = 120
target-version = "py311"
select = ["E", "F", "I", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-v --cov=app --cov-report=term-missing"
